libax25io: Link against zlib, if available.
So far only the presence of the zlib.h header was being tested and used if
present but the result was never linked against libz.  This means the
resulting libax25io library had no DT_NEEDED tag so linking a libax25io
built with zlib support into an application which itself wasn't being
linked against zlib would result in a link errror.  Whops.

In the dark past the issue must have been noticed but it was papered over
for the node package by linking against zlib, if present.  This might fail
if libax25io was built with zlib compression support but ax25-tools is
being built on a system without zlib.

Signed-off-by: Ralf Baechle <ralf@linux-mips.org>